better sample scripts
